+++ src/opts.rs
@@ -1,0 +1,36 @@
+use clap::{Parser, Subcommand};
+use std::path::Path;
+
+#[derive(Debug, Parser)]
+#[clap(name = "rcli", version, author, about, long_about)]
+pub struct Opts {
+    #[command(subcommand)]
+    pub cmd: SubCommand,
+}
+#[derive(Debug, Subcommand)]
+pub enum SubCommand {
+    #[clap(name = "csv", about = "Show csv, or convert to other formats")]
+    Csv(CsvOpts),
+}
+
+#[derive(Debug, Parser)]
+pub struct CsvOpts {
+    #[arg(short, long, value_parser=verify_input_file)]
+    pub input: String,
+    /// default_value默认值，传字符串然后由Parser convert
+    #[arg(short, long, default_value = "output.json")]
+    pub output: String,
+    #[arg(short, long, default_value_t = ',')]
+    pub delimiter: char,
+    /// default_value_t不需要转换直接写rust的类型
+    #[arg(long, default_value_t = true)]
+    pub header: bool,
+}
+
+fn verify_input_file(filename: &str) -> Result<String, &'static str> {
+    if Path::new(filename).exists() {
+        Ok(filename.to_string())
+    } else {
+        Err("File not found")
+    }
+}
